+++ app/login/page.tsx
@@ -1,7 +1,8 @@
 'use client';
 
 import { Icon_Logo } from '@/assets/icons';
-import KakaoLogin from '@/login/KakaoLogin';
+import { KakaoLogin } from '@/components/KakaoLogin';
+
 import Image from 'next/image';
 import { useEffect } from 'react';
 
@@ -12,13 +13,13 @@ export default function LoginPage() {
   }, []);
 
   return (
-    <div className="relative w-full min-h-screen">
+    <div className="relative min-h-screen w-full">
       <Image src="/login.webp" alt="메인 웨딩 사진" fill className="object-cover" priority />
-      <div className="absolute flex flex-col z-10 gap-[1.44rem] px-[2.3rem] pt-[19.2rem]">
+      <div className="absolute z-10 flex w-full flex-col gap-[1.44rem] px-[2.3rem] pt-[19.2rem]">
         <Icon_Logo width={142} height={23} />
-        <div className="w-[19.3rem] text-title2 font-bold text-common-0 z-10">딱 맞는 웨딩스냅 찾는 가장 쉬운 방법</div>
+        <div className="z-10 w-[19.3rem] text-title2 font-bold text-common-0">딱 맞는 웨딩스냅 찾는 가장 쉬운 방법</div>
       </div>
-      <div className="absolute bottom-[2rem] left-[2rem]">
+      <div className="absolute bottom-[2rem] w-full px-[2rem]">
         <KakaoLogin />
       </div>
     </div>
